Fix CodeQL orchestration import cycles and planner lambda
