+++ app/cli/interactive_shell/routing/handle_message_with_agent/orchestration/tools/synthetic_tool.py
@@ -13,7 +13,7 @@ from app.cli.interactive_shell.routing.handle_message_with_agent.orchestration.e
 from app.cli.interactive_shell.routing.handle_message_with_agent.orchestration.synthetic_scenarios import (
     list_rds_postgres_scenarios,
 )
-from app.cli.interactive_shell.routing.handle_message_with_agent.orchestration.tool_registry import (
+from app.cli.interactive_shell.routing.handle_message_with_agent.orchestration.tool_contracts import (
     ToolContext,
     ToolEntry,
     capability_not_explicitly_disabled,
