+++ app/cli/interactive_shell/routing/handle_message_with_agent/orchestration/tools/task_cancel_tool.py
@@ -15,7 +15,7 @@ from app.cli.interactive_shell.routing.handle_message_with_agent.orchestration.e
 from app.cli.interactive_shell.routing.handle_message_with_agent.orchestration.execution_tier import (
     ExecutionTier,
 )
-from app.cli.interactive_shell.routing.handle_message_with_agent.orchestration.tool_registry import (
+from app.cli.interactive_shell.routing.handle_message_with_agent.orchestration.tool_contracts import (
     ToolContext,
     ToolEntry,
     object_schema,
