+++ app/cli/interactive_shell/routing/tests/scenarios/follow_up/603-follow-up-last-investigation-wording.yml
@@ -1,0 +1,43 @@
+id: 603-follow-up-last-investigation-wording
+title: Follow-up last investigation wording
+intent_class: follow_up
+risk_level: medium
+input:
+  prompt: No, during the last investigation
+  surface: interactive_cli
+session:
+  has_prior_state: true
+  remote_connected: false
+  configured_integrations: []
+available_capabilities:
+  slash_commands: []
+  cli_commands: []
+  synthetic_suites: []
+notes: []
+
+route:
+  expected_kind: cli_agent
+  expected_signals: []
+  expected_command_text: null
+policy:
+  should_execute: false
+  has_unhandled_clause: true
+  fail_closed: true
+planned_actions:
+- kind: assistant_handoff
+  content: follow_up:last_investigation_summary
+  source: llm
+executed_actions: []
+response_contract:
+  must_contain_any:
+  - investigation
+  - follow-up
+  - alert text
+  must_not_contain:
+  - $ /
+history:
+  expected:
+  - type: cli_agent
+    ok: true
+tier: full
+runs: 3
